==> notebooks/01.00-maj-simple-logistic-regression.ipynb <==
# 01.00-maj-simple-logistic-regression

Mitchell Joseph
<br>Created: 2022-11-21</br>
Last Modified: 2022-11-21

In this notebook we explore a simple logistic regression model where we remap the output variables to 0 or 1 depending on whether or not a student got accepted to their top choice school.

import stan
# !pip install nest_asyncio
import nest_asyncio
nest_asyncio.apply()
import pandas as pd

import matplotlib.pyplot as plt
import numpy as np
import arviz as az
import seaborn as sns
# %load_ext rpy2.ipython

df_imp = pd.read_csv("../data/final_frame_imputed.csv")
df_nimp = pd.read_csv("../data/final_frame_non_imputed.csv")
