==> notebooks/01.02-maj-extract-DAG-features.ipynb <==
# 01.02-maj-extract-DAG-features

Extract values that are relevant to DAG.

import sqlite3
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

def run_query(db_path, query):
    """Query a database and return a pandas DataFrame."""
    conn = sqlite3.connect(db_path)
    df = pd.read_sql_query(query, conn)
    conn.close()
    return df

## High School

Relevant features
- GPA
- Habits of Mind
- Actions
- SAT scores

hs_path = "../data/00-external/HIGH_SCHOOL.db"

hs_query = """
SELECT [SubjectI.D.] as SUBJID,
    SATMATH,
    SATWriting,
    SATVerbal,
    CASE [Whatwasyouraveragegradeinhighschool?]
        WHEN "A or A+" THEN 4.0
        WHEN "A-" THEN 3.7
        WHEN "B+" THEN 3.3
        WHEN "B" THEN 3.0
        WHEN "B-" THEN 2.7
        WHEN "C+" THEN 2.3
        WHEN "C" THEN 2.0
        WHEN "D" THEN 1.0
    END AS GPA,
    CASE [HabitsofMind:Askquestionsinclass]
        WHEN "Frequently" THEN 2
        WHEN "Occasionally" THEN 1
        WHEN "Not at all" THEN 0
    END AS hab1,
    CASE [HabitsofMind:Supportyouropinionswithalogicalargument]
        WHEN "Frequently" THEN 2
        WHEN "Occasionally" THEN 1
        WHEN "Not at all" THEN 0
    END AS hab2,
    CASE [HabitsofMind:Seeksolutionstoproblemsandexplainthemtoothers]
        WHEN "Frequently" THEN 2
        WHEN "Occasionally" THEN 1
        WHEN "Not at all" THEN 0
    END as hab3,
    CASE [HabitsofMind:Reviseyourpaperstoimproveyourwriting]
        WHEN "Frequently" THEN 2
        WHEN "Occasionally" THEN 1
        WHEN "Not at all" THEN 0
    END as hab4
    FROM 'HIGH SCHOOL'
    WHERE SATMath IS NOT NULL
        AND SATWriting IS NOT NULL
        AND SATVerbal IS NOT NULL
        AND GPA IS NOT NULL
        AND hab1 IS NOT NULL
        AND hab2 IS NOT NULL
        AND hab3 IS NOT NULL
        AND hab4 IS NOT NULL
        AND Surveyyear == 2010;
"""
hs = run_query(hs_path, hs_query)
hs.dropna(inplace=True)

hs

## Demographics

demo_path = "../data/00-external/DEMOGRAPHICS.db"

query_demographics = """
SELECT [SubjectI.D.] as SUBJID,
    State AS state,
    [Yoursex:] AS gender,
    [Race/EthnicityGroup] AS race,
    ParentalIncome AS income
FROM DEMOGRAPHICS
WHERE Surveyyear == 2010
    AND state IS NOT NULL
    AND gender != 0
    AND race IS NOT NULL
    AND income IS NOT NULL;
"""

demo_df = run_query(demo_path, query_demographics)
demo_df.dropna(inplace=True)

# TFS

tfs_path = "../data/00-external/TFS_CHOICE_2008_2010.db"

    #    CASE WHEN ACCPT1ST = 2 THEN 1 ELSE 0 END as ACCPT1ST
query_tfs = """
SELECT SUBJID,
    CASE CHOICE
        WHEN 5 THEN -1
        WHEN 4 THEN -2
        WHEN 2 THEN -3
        WHEN 1 THEN -4
    END AS choice,
    CASE ACCPT1ST
        WHEN 1 THEN 0
        WHEN 2 THEN 1
    END AS accepted,
    ACERECODE as school,
    NUMAPPLY AS numapply
FROM CHOICE
WHERE Year == 2010
    AND CHOICE IS NOT NULL
    AND accepted IS NOT NULL
    AND school IS NOT NULL
    AND numapply IS NOT NULL;
"""

tfs= run_query(tfs_path, query_tfs)
tfs['choice'] = np.abs(tfs['choice'])
tfs.dropna(inplace=True)

tfs.dropna()

## Merging

print(hs.shape)
print(demo_df.shape)
print(tfs.shape)

merged_df = pd.merge(hs, demo_df, how='left', on="SUBJID")
print(merged_df.shape)
merged_df.isna().sum()

There were ~7k observations in the high school data frame that weren't in demographics

merged_df.dropna(inplace=True)
merged_df.shape

merged_df = pd.merge(merged_df, tfs, how='left', on='SUBJID')

merged_df.dropna(inplace=True)
print(merged_df.shape)

merged_df.to_csv("../data/01-raw/dag.csv", index=False)

merged_df
