==> notebooks/01.20-maj-clean-raw-data.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("../data/01-raw/dag.csv")
df

habit_cols = ['hab1', 'hab2', 'hab3', 'hab4']
df['habits'] = np.sum(df[habit_cols], axis=1)
df.drop(habit_cols, axis=1, inplace=True)

def parse_income(string):
    """Parse income string and return the average value."""
    try:
        lower, upper = "".join(string.replace("$", "")).replace(",","_").split('-')
        return np.mean([ int(lower[:2]), int(upper[:2])+1 ]) * 1_000
    except ValueError:
        if "or more" in string:
            return 250_000
        elif "or less" in string:
            return 10_000

# recode income
df['income'] = df['income'].apply(lambda x: parse_income(x))

def normalize_col(col):
    """Normalize a column so it has mean 0 and std 1."""
    m = np.mean(col)
    std = np.std(col)
    norm_vals = (col - m) / std

    assert np.round( np.mean(norm_vals), 5) == 0, "Mean is not 0"
    assert np.round( np.std(norm_vals), 5) == 1, "Std is not 1"

    return norm_vals

df

df['SAT'] = df['SATMath'] + df['SATVerbal'] + df['SATWriting']
cols_to_normalize = ["SATMath", "SATWriting", "SATVerbal", "GPA", "SAT"]
for col in cols_to_normalize:
    df[col] = normalize_col(df[col])

df['income'] = np.log10(df['income'])
df

df.to_csv("../data/02-processed/normalized_data.csv", index=False)
