==> notebooks/07.00-maj-random-forest.ipynb <==
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import classification_report
from sklearn.model_selection import train_test_split
from sklearn.model_selection import cross_val_score
from sklearn.metrics import f1_score
from sklearn.metrics import accuracy_score
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

from hyperopt import hp,fmin,tpe,STATUS_OK,Trials, space_eval

# Load the data
df1 = pd.read_csv("../data/01-raw/test_scores.csv")
df2 = pd.read_csv("../data/02-processed/normalized_data.csv")
df = pd.merge(df2, df1[["SUBJID", "State"]], how='left', on="SUBJID")
df.dropna(inplace=True)

# Clean the data
feature_cols = ["SAT", "GPA", "income"]
outcome_col = ["accepted"]

# Get dummy variables for school
df["school"] = df["school"].astype(np.int64)
dummy_school = pd.get_dummies(df["school"], prefix="school")
dummy_state = pd.get_dummies(df["State"], prefix="", prefix_sep="")
dummy_race = pd.get_dummies(df["race"], prefix="", prefix_sep="")
dummy_gender = pd.get_dummies(df["gender"], prefix="", prefix_sep="")
col_labels = feature_cols + list(dummy_school.columns) + list(dummy_state.columns) + list(dummy_race.columns) + list(dummy_gender.columns)
X = pd.concat([df[feature_cols], dummy_school, dummy_state, dummy_race, dummy_gender], axis=1).copy()

X_train_df, X_test_df, y_train_df, y_test_df = train_test_split(
    X,
    df[outcome_col],
    random_state=17)

# Convert dataframes to np arrays
X_train = X_train_df.copy().to_numpy()
X_test  = X_test_df.copy().to_numpy()
y_train = y_train_df.copy().astype(np.int8).to_numpy().flatten()
y_test = y_test_df.copy().astype(np.int8).to_numpy().flatten()

rf = RandomForestClassifier(n_estimators=1_000,
                            n_jobs=-1,
                            random_state=17)
rf.fit(X_train, y_train)

y_pred = rf.predict(X_test)

print(classification_report(y_test, y_pred))

## Let's try with the imputed data from Chris

df = pd.read_csv("../data/final_frame_non_imputed.csv")
X = df.drop(["CHOICE", "ACCPT1ST", "Unnamed: 0"], axis=1)
Y = df["ACCPT1ST"].replace({1:0, 2:1})
X_train_df, X_test_df, y_train_df, y_test_df = train_test_split(
    X,
    Y,
    random_state=17)
# Convert dataframes to np arrays
X_train = X_train_df.copy().to_numpy()
X_test  = X_test_df.copy().to_numpy()
y_train = y_train_df.copy().astype(np.int8).to_numpy().flatten()
y_test = y_test_df.copy().astype(np.int8).to_numpy().flatten()

rf = RandomForestClassifier(n_jobs=-1)
rf.fit(X_train, y_train)

y_pred = rf.predict(X_test)

print(classification_report(y_test, y_pred))

## Bayesian Optimization

space = dict(
    # n_estimators = hp.choice("n_estimators", [10, 50, 100, 500, 1200]),
    n_estimators = hp.randint("n_estimators", 2, 100),
    criterion = hp.choice("criterion", ["entropy", "gini", "log_loss"]), 
    max_depth = hp.randint("max_depth", 10, 5000),
    min_samples_split = hp.uniform("min_samples_split", 0, 1),
    min_samples_leaf = hp.uniform("min_samples_leaf", 0, 1),
    max_features = hp.choice("max_features", ["sqrt", "log2", None])
)

def objective(space):
    model = RandomForestClassifier(
        n_estimators=space["n_estimators"],
        criterion=space["criterion"],
        max_depth=space["max_depth"],
        min_samples_split=space["min_samples_split"],
        min_samples_leaf=space["min_samples_leaf"],
        max_features=space["max_features"],
        n_jobs=-1
    )
    model.fit(X_train, y_train)
    y_pred = model.predict(X_test)
    f1 = f1_score(y_test, y_pred, average="macro")
    return {"loss": -f1, "status": STATUS_OK}
    # accuracy = accuracy_score(y_test, y_pred)
    # return {"loss": -accuracy, "status": STATUS_OK}

best_params = fmin(
    fn=objective,
    space=space,
    algo=tpe.suggest,  # Tree of Parzen Estimators (Bayesian approach)
    max_evals=100
)

space_eval(space, best_params)

rf = RandomForestClassifier(**space_eval(space, best_params))
rf.fit(X_train, y_train)

y_pred = rf.predict(X_test)
print(classification_report(y_test, y_pred))
