==> notebooks/09.00-maj-causal-inference-direct-effects.ipynb <==
# 09.00-maj-causal-inference-direct-effects

In this notebook we revisit our causal inference model once again.
Using the online toolkit from Daggity, we can measure both the total effect and direct effect of

- SAT scores on acceptance
- Race on acceptance
- Gender on acceptance
<br></br>
- Total effect of SAT: $Y \sim SAT + GPA + Income$
- Direct effect of SAT: $Y \sim SAT + Gender + Race + School$
- Total effect of Gender/Race: $Y \sim Gender + GPA + SAT + Race + School$
- Direct effect of Gender/Race: $Y \sim Gender/Race$

import pandas as pd
import numpy as np
import arviz as az
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report

from psis import *

import stan
import nest_asyncio
nest_asyncio.apply()

logistic_regression_code = """
/**
 * Logistic regression t-prior
 *
 * Priors:
 *     weights - student t
 *     intercept - student t
 */
data {
    int n;                        // number of data points
    int d;                        // explanatory variable dimension
    matrix[n, d] X;               // explanatory variable
    array[n] int <lower=0, upper=1> y;  // response variable
    int<lower=1> p_alpha_df;      // prior degrees of freedom for alpha
    real p_alpha_loc;             // prior location for alpha
    real p_alpha_scale;           // prior scale for alpha
    int<lower=1> p_beta_df;       // prior degrees of freedom for beta
    real p_beta_loc;     // prior location for beta
    real p_beta_scale;            // prior scale for beta

    int<lower=0> N_new;
    matrix[N_new, d] X_new;
}
parameters {
    real alpha;      // intercept
    vector[d] beta;  // explanatory variable weights
}
transformed parameters {
    vector[n] eta;  // linear predictor
    eta = alpha + X * beta;
}
model {
    alpha ~ student_t(p_alpha_df, p_alpha_loc, p_alpha_scale);
    beta ~ student_t(p_beta_df, p_beta_loc, p_beta_scale);
    y ~ bernoulli_logit(eta);
}
generated quantities {
    // calculate the log likelihood
    vector[n] log_lik;
    for (i in 1:n)
        log_lik[i] = bernoulli_logit_lpmf(y[i] | eta[i]);

    // generate values for Y tilde
    array[n] int<lower=0, upper=1> Y_tilde;
    Y_tilde = bernoulli_logit_rng(eta);

    // Make predictions
    vector[N_new] y_new;
    for (j in 1:N_new) {
        y_new[j] = bernoulli_logit_rng(alpha + X_new[j] * beta);
    }
}
"""



def run_model(data_dict):
    logistic_regression = stan.build(logistic_regression_code, data=data_dict)
    fit = logistic_regression.sample()
    return fit

def get_summary_table(fit, X_train_df, col=None):
    az_fit = az.from_pystan(fit)
    summary_table = az.summary(az_fit, var_names=['alpha', 'beta'])
    index_labels = ["intercept"] + list(X_train_df.columns)
    summary_table.index = index_labels
    if col:
        return summary_table.loc[col, :]
    else:
        return summary_table

df = pd.read_csv("../data/02-processed/normalized_data.csv")
df.dropna(inplace=True)
df

## Total effect of SAT on Acceptance

total_sat_cols = ["SAT", "GPA", "income"]
total_gender_cols = ["gender", "GPA", "SAT", "race", "school"]
outcome = ["accepted"]

X = df[total_sat_cols].copy()
Y = df[outcome].copy()

X_train_df, X_test_df, y_train_df, y_test_df = train_test_split(
    X,
    Y,
    train_size=10_000,
    test_size=10,
    random_state=17)

# Convert dataframes to np arrays
X_train = X_train_df.copy().to_numpy()
X_test  = X_test_df.copy().to_numpy()
y_train = y_train_df.copy().astype(np.int8).to_numpy().flatten()
y_test = y_test_df.copy().astype(np.int8).to_numpy().flatten()

# Initialize data dictionary
n = X_train.shape[0]
d = X_train.shape[1]
degf = 1
data_dict = dict(
    n=n,                # num data points
    d=d,                # num features
    X=X_train,                # data matrix  
    y=y_train,                # response variable
    p_alpha_df=degf,   # prior deg freedom for alpha
    p_alpha_loc=0,      # prior location for alpha
    p_alpha_scale=5,  # prior scale for alpha
    p_beta_df=degf,    # prior deg freedom for beta
    p_beta_loc=0,       # prior location for beta
    p_beta_scale=5,    # prior scale for beta
    N_new=X_test.shape[0],
    X_new=X_test
)

fit = run_model(data_dict)

get_summary_table(fit, X_train_df)

The total effect of SAT scores on acceptance is $\exp(-0.372) \approx 0.69 \approx 31\%$  decrease in acceptance.

## Direct effect of SAT scores on Acceptance

direct_sat_cols = ["SAT", "gender", "race", "school"]
dummy_school = pd.get_dummies(df["school"], prefix="school")
dummy_race = pd.get_dummies(df["race"], prefix="", prefix_sep="")
dummy_gender = pd.get_dummies(df["gender"], prefix="", prefix_sep="")
X = pd.concat([df["SAT"], dummy_school, dummy_race, dummy_gender], axis=1).copy()
Y = df[outcome].copy()

X_train_df, X_test_df, y_train_df, y_test_df = train_test_split(
    X,
    Y,
    train_size=10_000,
    test_size=10,
    random_state=17)

# Convert dataframes to np arrays
X_train = X_train_df.copy().to_numpy()
X_test  = X_test_df.copy().to_numpy()
y_train = y_train_df.copy().astype(np.int8).to_numpy().flatten()
y_test = y_test_df.copy().astype(np.int8).to_numpy().flatten()

# Initialize data dictionary
n = X_train.shape[0]
d = X_train.shape[1]
degf = 7
data_dict = dict(
    n=n,                # num data points
    d=d,                # num features
    X=X_train,                # data matrix  
    y=y_train,                # response variable
    p_alpha_df=degf,   # prior deg freedom for alpha
    p_alpha_loc=0,      # prior location for alpha
    p_alpha_scale=5,  # prior scale for alpha
    p_beta_df=degf,    # prior deg freedom for beta
    p_beta_loc=0,       # prior location for beta
    p_beta_scale=5,    # prior scale for beta
    N_new=X_test.shape[0],
    X_new=X_test
)

fit = run_model(data_dict)

get_summary_table(fit, X_train_df, ["SAT"])

## Total effect of Gender/Race on Acceptance

X = pd.concat([df["SAT"], dummy_gender, dummy_race, dummy_school], axis=1)

X_train_df, X_test_df, y_train_df, y_test_df = train_test_split(
    X,
    Y,
    train_size=10_000,
    test_size=10,
    random_state=17)

# Convert dataframes to np arrays
X_train = X_train_df.copy().to_numpy()
X_test  = X_test_df.copy().to_numpy()
y_train = y_train_df.copy().astype(np.int8).to_numpy().flatten()
y_test = y_test_df.copy().astype(np.int8).to_numpy().flatten()

# Initialize data dictionary
n = X_train.shape[0]
d = X_train.shape[1]
degf = 7
data_dict = dict(
    n=n,                # num data points
    d=d,                # num features
    X=X_train,                # data matrix  
    y=y_train,                # response variable
    p_alpha_df=degf,   # prior deg freedom for alpha
    p_alpha_loc=0,      # prior location for alpha
    p_alpha_scale=5,  # prior scale for alpha
    p_beta_df=degf,    # prior deg freedom for beta
    p_beta_loc=0,       # prior location for beta
    p_beta_scale=5,    # prior scale for beta
    N_new=X_test.shape[0],
    X_new=X_test
)

%%capture
fit = run_model(data_dict)

get_summary_table(fit, X_train_df, list(dummy_gender) + list(dummy_race))

Both gender and race have a positive total effect on acceptance with the exception of being Asian and Hispanic. 
These correspond to a $\exp(-0.718) \approx 0.49 \approx 51\%$ and $\exp(-0.307) \approx 0.74 \approx 26\%$ reduction in the odds of getting accepted into their first choice school.
However the hdi for all variables includes negative and positive values and so the mean being centered between 0 to 1 for all variables could indicate a low degree of confidence for these estimated values. 

## Direct effect of Race on Accpetance

X = dummy_race

X_train_df, X_test_df, y_train_df, y_test_df = train_test_split(
    X,
    Y,
    train_size=10_000,
    test_size=10,
    random_state=17)

# Convert dataframes to np arrays
X_train = X_train_df.copy().to_numpy()
X_test  = X_test_df.copy().to_numpy()
y_train = y_train_df.copy().astype(np.int8).to_numpy().flatten()
y_test = y_test_df.copy().astype(np.int8).to_numpy().flatten()

# Initialize data dictionary
n = X_train.shape[0]
d = X_train.shape[1]
degf = 7
data_dict = dict(
    n=n,                # num data points
    d=d,                # num features
    X=X_train,                # data matrix  
    y=y_train,                # response variable
    p_alpha_df=degf,   # prior deg freedom for alpha
    p_alpha_loc=0,      # prior location for alpha
    p_alpha_scale=5,  # prior scale for alpha
    p_beta_df=degf,    # prior deg freedom for beta
    p_beta_loc=0,       # prior location for beta
    p_beta_scale=5,    # prior scale for beta
    N_new=X_test.shape[0],
    X_new=X_test
)

%%capture
fit = run_model(data_dict)

get_summary_table(fit, X_train_df)

The direct effect of being Asian corresponds to a 56\% reduction in the odds of being accepted. 
Being hispanic now corresponds to only a $.002%$ reduction in the odds of being accepted.

## Direct effect of Gender on Acceptance

X = dummy_gender

X_train_df, X_test_df, y_train_df, y_test_df = train_test_split(
    X,
    Y,
    train_size=10_000,
    test_size=10,
    random_state=17)

# Convert dataframes to np arrays
X_train = X_train_df.copy().to_numpy()
X_test  = X_test_df.copy().to_numpy()
y_train = y_train_df.copy().astype(np.int8).to_numpy().flatten()
y_test = y_test_df.copy().astype(np.int8).to_numpy().flatten()

# Initialize data dictionary
n = X_train.shape[0]
d = X_train.shape[1]
degf = 7
data_dict = dict(
    n=n,                # num data points
    d=d,                # num features
    X=X_train,                # data matrix  
    y=y_train,                # response variable
    p_alpha_df=degf,   # prior deg freedom for alpha
    p_alpha_loc=0,      # prior location for alpha
    p_alpha_scale=5,  # prior scale for alpha
    p_beta_df=degf,    # prior deg freedom for beta
    p_beta_loc=0,       # prior location for beta
    p_beta_scale=5,    # prior scale for beta
    N_new=X_test.shape[0],
    X_new=X_test
)

%%capture
fit = run_model(data_dict)

get_summary_table(fit, X_train_df)

Both male and females have a po
